+++ packages/swapkit/api/src/thorswapApiV2/endpoints.ts
@@ -1,11 +1,17 @@
 import {
+  type ProviderName,
   type QuoteRequest as QuoteRequestV2,
   type QuoteResponse,
   QuoteResponseSchema,
   RequestClient,
   SwapKitError,
 } from "@swapkit/helpers";
-import type { TrackerParams, TrackerResponse } from "./types";
+import type {
+  TokenListProvidersResponse,
+  TokensResponseV2,
+  TrackerParams,
+  TrackerResponse,
+} from "./types";
 
 const baseUrl = "https://api.swapkit.dev";
 const baseUrlDev = "https://dev-api.swapkit.dev";
@@ -27,4 +33,12 @@ export async function getSwapQuoteV2(searchParams: QuoteRequestV2, isDev = false
   } catch (error) {
     throw new SwapKitError("api_v2_invalid_response", error);
   }
+}
+
+export function getTokenListProvidersV2() {
+  return RequestClient.get<TokenListProvidersResponse>(`${baseUrl}/providers`);
+}
+
+export function getTokenListV2(provider: ProviderName) {
+  return RequestClient.get<TokensResponseV2>(`${baseUrl}/tokens?provider=${provider}`);
 }