+++ packages/swapkit/api/src/thorswapStatic/endpoints.ts
@@ -1,8 +1,8 @@
-import { RequestClient } from "@swapkit/helpers";
+import { AssetValue, RequestClient, getChainIdentifier } from "@swapkit/helpers";
 
 import type { TokensResponse } from "./types.ts";
 
-const baseUrl = "https://static.thorswap.finance";
+const baseUrl = "https://static.thorswap.net";
 
 export function getTokenList(tokenListName: string) {
   return RequestClient.get<TokensResponse>(`${baseUrl}/token-list/${tokenListName}.json`);
@@ -10,4 +10,10 @@ export function getTokenList(tokenListName: string) {
 
 export function getLogoForAsset(assetString: string) {
   return `${baseUrl}/token-list/images/${assetString.toLowerCase()}.png`;
+}
+
+export function getChainLogoForAsset(assetString: string) {
+  const { chain } = AssetValue.fromStringSync(assetString);
+  const chainIdentifier = getChainIdentifier(chain).toLowerCase();
+  return `${baseUrl}/token-list/images/${chainIdentifier}.png`;
 }