+++ packages/swapkit/chainflip/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @swapkit/chainflip
 
+## 1.0.0-rc.65
+
+### Patch Changes
+
+- Updated dependencies [[`dd7c87e`](https://github.com/thorswap/SwapKit/commit/dd7c87e457fcc275d34c5aee0a08ebf3a9d47612)]:
+  - @swapkit/helpers@1.0.0-rc.110
+  - @swapkit/api@1.0.0-rc.78
+  - @swapkit/toolbox-evm@1.0.0-rc.116
+  - @swapkit/toolbox-substrate@1.0.0-rc.45
+
+## 1.0.0-rc.64
+
+### Patch Changes
+
+- Updated dependencies [[`aa10e7a`](https://github.com/thorswap/SwapKit/commit/aa10e7a847c18abacbbdbefec8a93487cff2a524)]:
+  - @swapkit/helpers@1.0.0-rc.109
+  - @swapkit/api@1.0.0-rc.77
+  - @swapkit/toolbox-evm@1.0.0-rc.115
+  - @swapkit/toolbox-substrate@1.0.0-rc.44
+
+## 1.0.0-rc.63
+
+### Patch Changes
+
+- Updated dependencies [[`f860028`](https://github.com/thorswap/SwapKit/commit/f860028cf61f4818d74b57345ae643d6af07007e)]:
+  - @swapkit/helpers@1.0.0-rc.108
+  - @swapkit/api@1.0.0-rc.76
+  - @swapkit/toolbox-evm@1.0.0-rc.114
+  - @swapkit/toolbox-substrate@1.0.0-rc.43
+
+## 1.0.0-rc.62
+
+### Patch Changes
+
+- Updated dependencies [[`e72a9b3`](https://github.com/thorswap/SwapKit/commit/e72a9b3d8e6874175a3c1871a0b7528fb249cd18)]:
+  - @swapkit/api@1.0.0-rc.75
+  - @swapkit/helpers@1.0.0-rc.107
+  - @swapkit/toolbox-evm@1.0.0-rc.113
+  - @swapkit/toolbox-substrate@1.0.0-rc.42
+
+## 1.0.0-rc.61
+
+### Patch Changes
+
+- Updated dependencies [[`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89)]:
+  - @swapkit/helpers@1.0.0-rc.106
+  - @swapkit/api@1.0.0-rc.74
+  - @swapkit/toolbox-evm@1.0.0-rc.112
+  - @swapkit/toolbox-substrate@1.0.0-rc.41
+
+## 1.0.0-rc.60
+
+### Patch Changes
+
+- Updated dependencies [[`aa93dc5`](https://github.com/thorswap/SwapKit/commit/aa93dc509e6b8e0c3cf2615d072ab7f059784c5c)]:
+  - @swapkit/helpers@1.0.0-rc.105
+  - @swapkit/api@1.0.0-rc.73
+  - @swapkit/toolbox-evm@1.0.0-rc.111
+  - @swapkit/toolbox-substrate@1.0.0-rc.40
+
 ## 1.0.0-rc.59
 
 ### Minor Changes