+++ packages/wallets/ledger/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @swapkit/wallet-ledger
 
+## 1.0.0-rc.158
+
+### Patch Changes
+
+- Updated dependencies [[`dd7c87e`](https://github.com/thorswap/SwapKit/commit/dd7c87e457fcc275d34c5aee0a08ebf3a9d47612)]:
+  - @swapkit/helpers@1.0.0-rc.110
+  - @swapkit/toolbox-cosmos@1.0.0-rc.136
+  - @swapkit/toolbox-evm@1.0.0-rc.116
+  - @swapkit/toolbox-utxo@1.0.0-rc.119
+
+## 1.0.0-rc.157
+
+### Patch Changes
+
+- Updated dependencies [[`aa10e7a`](https://github.com/thorswap/SwapKit/commit/aa10e7a847c18abacbbdbefec8a93487cff2a524)]:
+  - @swapkit/helpers@1.0.0-rc.109
+  - @swapkit/toolbox-cosmos@1.0.0-rc.135
+  - @swapkit/toolbox-evm@1.0.0-rc.115
+  - @swapkit/toolbox-utxo@1.0.0-rc.118
+
+## 1.0.0-rc.156
+
+### Patch Changes
+
+- Updated dependencies [[`f860028`](https://github.com/thorswap/SwapKit/commit/f860028cf61f4818d74b57345ae643d6af07007e)]:
+  - @swapkit/helpers@1.0.0-rc.108
+  - @swapkit/toolbox-cosmos@1.0.0-rc.134
+  - @swapkit/toolbox-evm@1.0.0-rc.114
+  - @swapkit/toolbox-utxo@1.0.0-rc.117
+
+## 1.0.0-rc.155
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@1.0.0-rc.107
+  - @swapkit/toolbox-cosmos@1.0.0-rc.133
+  - @swapkit/toolbox-evm@1.0.0-rc.113
+  - @swapkit/toolbox-utxo@1.0.0-rc.116
+
+## 1.0.0-rc.154
+
+### Patch Changes
+
+- Updated dependencies [[`f389990`](https://github.com/thorswap/SwapKit/commit/f389990a4a14566315fe353a464d916fc247ce89)]:
+  - @swapkit/helpers@1.0.0-rc.106
+  - @swapkit/toolbox-cosmos@1.0.0-rc.132
+  - @swapkit/toolbox-evm@1.0.0-rc.112
+  - @swapkit/toolbox-utxo@1.0.0-rc.115
+
+## 1.0.0-rc.153
+
+### Patch Changes
+
+- Updated dependencies [[`aa93dc5`](https://github.com/thorswap/SwapKit/commit/aa93dc509e6b8e0c3cf2615d072ab7f059784c5c)]:
+  - @swapkit/helpers@1.0.0-rc.105
+  - @swapkit/toolbox-cosmos@1.0.0-rc.131
+  - @swapkit/toolbox-evm@1.0.0-rc.111
+  - @swapkit/toolbox-utxo@1.0.0-rc.114
+
 ## 1.0.0-rc.152
 
 ### Patch Changes